+++ src/main/java/com/teamddd/duckmap/dto/review/CreateReviewReq.java
@@ -1,4 +1,4 @@
-package com.teamddd.duckmap.dto;
+package com.teamddd.duckmap.dto.review;
 
 import java.util.List;
 
@@ -7,11 +7,15 @@ import javax.validation.constraints.Size;
 
 import org.hibernate.validator.constraints.Range;
 
+import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.Getter;
 
 @Getter
-public class UpdateReviewReq {
+public class CreateReviewReq {
+	@NotNull
+	private Long eventId;
 	@Range(min = 0, max = 5)
+	@Schema(defaultValue = "0")
 	private int score;
 	private String content;
 	@NotNull
