+++ src/main/java/com/teamddd/duckmap/dto/review/ReviewRes.java
@@ -1,7 +1,9 @@
-package com.teamddd.duckmap.dto;
+package com.teamddd.duckmap.dto.review;
 
 import java.time.LocalDateTime;
 import java.util.List;
+
+import com.teamddd.duckmap.dto.ImageRes;
 
 import lombok.Builder;
 import lombok.Getter;
