+++ src/main/java/com/teamddd/duckmap/dto/review/ReviewSearchParam.java
@@ -1,4 +1,4 @@
-package com.teamddd.duckmap.dto;
+package com.teamddd.duckmap.dto.review;
 
 import lombok.Builder;
 import lombok.Getter;
