+++ src/main/java/com/teamddd/duckmap/dto/review/ReviewsRes.java
@@ -1,4 +1,6 @@
-package com.teamddd.duckmap.dto;
+package com.teamddd.duckmap.dto.review;
+
+import com.teamddd.duckmap.dto.ImageRes;
 
 import lombok.Builder;
 import lombok.Getter;
