+++ src/main/java/com/teamddd/duckmap/dto/review/UpdateReviewReq.java
@@ -1,4 +1,4 @@
-package com.teamddd.duckmap.dto;
+package com.teamddd.duckmap.dto.review;
 
 import java.util.List;
 
@@ -7,15 +7,11 @@ import javax.validation.constraints.Size;
 
 import org.hibernate.validator.constraints.Range;
 
-import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.Getter;
 
 @Getter
-public class CreateReviewReq {
-	@NotNull
-	private Long eventId;
+public class UpdateReviewReq {
 	@Range(min = 0, max = 5)
-	@Schema(defaultValue = "0")
 	private int score;
 	private String content;
 	@NotNull
